feat: 지원 대상 박스
