+++ src/@pages/recruit/style.jsx
@@ -1,90 +1,58 @@
 import styled from 'styled-components';
-import { Button } from 'reactstrap';
 
-export const HeaderContainer = styled.div`
+export const RecuritBody = styled.div`
+  width: 70%;
   display: flex;
-  justify-content: center;
-  align-items: center;
   flex-direction: column;
-  /* border: 1px solid black; */
+  justify-content: start;
+  align-items: center;
+  border: 1px solid blue;
+`;
+export const TargetContainer = styled.div`
   width: 100%;
-  background: var(--light-gray-gray-050, #f7f7fa);
-  color: #000;
-`;
-
-export const HeaderTitleWrapper = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
   flex-direction: column;
-  margin-top: 6.25rem;
-  margin-bottom: 8.25rem;
+  border: 1px solid red;
+  margin-top: 8.56rem;
+  margin-bottom: 9.06rem;
 `;
 
-export const RowPosters = styled.div`
-  display: flex;
-  overflow-y: hidden;
-  overflow-x: scroll;
-  scroll-behavior: smooth;
-  /* z-index: 500; */
-`;
-
-export const HeaderTitleText = styled.h2`
+export const TargetTitle = styled.h2`
   color: #000;
-  text-align: center;
+  font-feature-settings: 'clig' off, 'liga' off;
+  margin-bottom: 2.25rem;
+`;
+
+export const TargetDetail = styled.h5`
+  color: var(--light-gray-gray-900, #2b2d36);
   font-feature-settings: 'clig' off, 'liga' off;
   margin-bottom: 1.5rem;
 `;
 
-export const HeaderUnivContainer = styled.div`
+export const TargetRowDetailWrapper = styled.div`
   display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-direction: column;
-  margin-bottom: 4.5rem;
-  width: 100%;
-`;
-
-export const HeaderUnivTitleText = styled.h5`
-  color: #000;
-  font-feature-settings: 'clig' off, 'liga' off;
-  margin-bottom: 1.5rem;
-`;
-
-export const HeaderUnivListContainer = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: center;
   flex-direction: row;
-  /* border: 1px solid purple; */
-  width: 100%;
+  gap: 1rem;
 `;
 
-export const GoormBtn = styled(Button)`
-  display: inline-flex;
-  padding: 0.875rem 1.5rem;
-  flex-direction: column;
-  justify-content: center;
+export const TargetRowDetailBox = styled.div`
+  display: flex;
+  flex-direction: row;
   align-items: center;
-  gap: 0.625rem;
-  height: 3rem;
-  flex-shrink: 0;
-  background-color: #5094fa;
+  justify-content: start;
+  gap: 0.5rem;
   border-radius: 0.5rem;
-  background: var(--light-blue-blue-500-primary, #5094fa);
-  color: var(--light-gray-gray-000, #fff);
-  text-align: center;
+  padding: 0.75rem 1rem;
+  background: var(--light-gray-gray-100, #f0f0f5);
+`;
+
+export const TargetDetailBoxText = styled.p`
+  color: #000;
   font-feature-settings: 'clig' off, 'liga' off;
-  /* KOR/H6ㅣ1rem */
+
+  /* KOR/subtitle-1ㅣ0.875rem */
   font-family: Pretendard;
-  font-size: 1rem;
+  font-size: 0.875rem;
   font-style: normal;
   font-weight: 500;
-  line-height: 1.5rem; /* 150% */
-  letter-spacing: -0.00625rem;
-`;
-
-export const BtnText = styled.h6`
-  text-align: center;
-  font-feature-settings: 'clig' off, 'liga' off;
+  margin-bottom: 0;
 `;
